api/inventory-service/requirements.txt
fastapi
pydantic
uvicorn
tortoise-orm